Check for sentinel version control files/dirs inside working dirs, to avoid errors
A problem existed prior to this commit: if a package was changed from
darcs to git, then running "git pull" in the package's working dir
would silently perform that git operation in the root of the Melpa
repo. Checking that the version control for a working dir is as
expected prevents this issue.