Added pb/checkout-raw to obtain files from raw http urls or filesystem.
Many emacs-lisp packages are pretty tiny and contain only a single
.el file, some are exposed as plain files over http.
The raw fetcher can download any http url file or obtain it from
the local filesystem.

For single file packages hosted on github, I'd prefer to create a
recipe with the url of its raw .el content instead of using git to
download it.