Web: link to github and emacswiki
Adds an additional column to the table on http://melpa.milkbox.net
which provides a link directly to the github or emacswiki page for
a recipe.

My main gripe with using both `package-list-packages` and the web
UI is the extra effort required just to go check out a project's
README or similar. This should now be easier for the vast majority
of recipes in MELPA.

For emacswiki pages, only those recipes whose file matches the name
of the recipe are linked (ie, if the recipe uses :files, then no
link will be generated). This is Not Perfect(tm), but not bad either.