Add minitest-mode
This is a minitest minor mode. This mode is pretty similiar to the rspec
mode, however it is for minitest. I am using it in the daily basis to
develop ruby apps such as rails, ams, and Shopify projects.

I am the creator of it, and maintener.

Link for it: https://github.com/arthurnn/minitest-emacs