Add timestamp versions to packages with no tags.
Building packages for MELPA stable where there is no tag yet will
not cause an error, but instead try to guess the last official
version number from the file header or pkg.el file, and append a
timestamp to that.

This way, packages which would otherwise not be in MELPA stable
at all now are available via that repository as well, while authors
can still declare stable versions and have those be used instead.