Adding a recipe for bison-mode.
This is a major mode for Bison grammars. I'm not the original author,
the original author wrote the major mode in 1998.

I have created a git repository and updated the major mode to conform
more closely to modern elisp conventions. I've fixed the worst build
errors, but I'm aware that checkdoc and flycheck still find plenty of
things wrong with the code.