noccur: Run multi-occur on project/dired files
occur-mode is one of the awesome modes that come builtin with Emacs.
Sometimes I just want to run multi-occur on all (or a subdirectory) of a
project I'm working on. Used with keyboard macros it makes it a snap to
perform modifications on many buffers at once.

https://github.com/NicolasPetton/noccur.el

I'm a friend of the maintainer and contributed patches.