Updating official repository for ein recipe
The original author of emacs-ipython-notebook, @tkf, has been inactive
on the project for nine months or so. Given that the project is
currently incompatible with current versions of IPython, I have forked
the project to millejoh/emacs-ipython-notebook to provide this
compatibility.

This change updates the MELPA recipe for ein to point to my repository,
since I appear to have become the de-facto maintainer of the project.