Add recipe for aggressive-fill-paragraph package
An emacs minor-mode for keeping paragraphs filled in both comments and
prose.

Each time a space is inserted the current paragraph is refilled. Some
special behaviour is implemented in a few major modes where the default
fill-paragraph behaviour would affect code as well as comments.

The package repository is
[here](https://github.com/davidshepherd7/aggressive-fill-paragraph-mode).

I am the author of the package.