A better smart-assignment for R (and S) with the =-key, no underscores.
This provides a minor mode within ESS for R and S source and interaction
that makes the =-key intelligently select the intended meaning based
on the previous context.

The package repository is on github at
https://github.com/genovese/ess-smart-equals.
I am the original package author and maintainer.

This has been checked with flyspell-package and by building
in a clone of the MELPA repository, testing under a sandbox,
as described in the MELPA documentation.

Note that flycheck-package cannot determine that ess-site is the
proper feature to load for package ess and so signals a problem.
But there is no problem, and the package loads correctly.