Change package named electric-spacing
The current "electric-spacing" is a package started in 2014, while there
already was a much older and widely used package called "smart-operator"
which was renamed to "electric-spacing" in 2012.

The classic "electric-spacing" is in GNU ELPA and I think that MELPA
should revert to this package and the newer one should be renamed.