explicitly set branch for magit and related packages
I have just released 1.4.0 and am now preparing to release 2.1.0.  I am
now making the `next' branches of the various repositories the HEAD, but
would like the Melpa packages to be build from the `master' branches for
a little longer.  I will switch to the `next' branches approximately to
weeks before 2.1.0 is actually released.