Adjust circe to the new upstream repository layout.
Remove lui, lcs and shorten, as they are no longer separate
packages, but simply included in circe. This should not pose
a problem, as no other packages in MELPA depend on them.

As the tracking package is used by another package, leave the
recipe, but adjust it to include the depended-on shorten.el.