Generalize regexp for tags to identify stable versions
This allows for tags that begin with the name of the package followed by
a version number to be identified as tags corresponding to stable
versions. E.g. "foo-mode-v.1.0.0" will now be identified.

The use case is when multiple packages are kept in a single
repo (which is admitted sub-optimal) or if the package is kept in a repo
with other code.