Update the build scripts to work from within repo.
Attempt to cleanup scripts a bit so that everything is used within the
repo where it can be. Most importantly, we will not sync files anymore
with the 'www' directory and rather just serve files from the repo
itself. The json and html files can be built in-place then and we should
avoid some weird syncing problems.

This should also make it simpler to see what files are doing which
things.