Add bind-map
bind-map is a macro that generalizes the leader concept from vim for
emacs users. It is a simple way to bind a keymap to multiple prefixes
with the option to condition on the current major mode, minor mode, or
evil state. It should be an easy way to organize key bindings in
personal keymaps.