remove unmaintained starter-kit
The original and very influential Starter-Kit has come to age and should
be considered unmaintained.

See https://github.com/magit/magit/issues/1923#issuecomment-116082321.

The Starter-Kit was also used as an example in the README.md, so that
also has to be replaced.  Projectile seems to be a good choice for that,
better than the Starter-Kit actually, because it is more typical.