remove outdated fork of built-in cperl-mode
This is an outdated fork of the built-in package "cperl-mode".  It
should be removed so that users do not end up installing an older
version when they try to install a newer version than what is
distributed with Emacs.

The commentary of the fork mentions that it "contains support for the
syntax added by the MooseX::Declare CPAN module, as well as Perl 5.10
keyword support".  I do not know whether the built-in version also
supports these things too now.  If not, then an attempt should be made
to get these changes merged upstream.  The fork has not incorporated
any upstream changes during the last two years.