Add worse-defaults
Things in this package represent better (worse?) default settings for
many things that Emacs ships with.

This defaults are also somewhat opinionated, which might not be
acceptable for everyone, hence worse in the title.

In short, a lot. This is definitely not one of those minimalist packages
that try to find the golden middle between sensible and practical, use
better-defaults if you prefer something like that.

You should really take a look at the source file on GitHub and read the
comments, there are plenty of them.

A convenience function goto-matching-paren is also provided, it is bound
to `C-c {`.