Local declarative programming with Org elements.
org-dp (https://github.com/tj64/org-dp) is meant for local programming,
i.e. for situations where the complete parse tree of an Org-mode file is
not needed, but the focus is on the Org element at point.

It leaves parsing and interpreting to the Org-mode parser framework and
creates, updates and deletes Org elements by acting on their uniform
internal (p-list) representations rather than on their diverse textual
representations.

I'm the library's only author and maintainer.