Move package-build.el to a seperate repository and merge it as a subtree (#4121)
* Move package-build.el to a separate repository

This commit removes "package-build.el" and adjusts "Makefile" to expect
that file to be located in "package-build/" instead of at the top-level.

This commit also adjusts the documentation and adds a new make target
"pull-package-build" which performs a subtree merge after fetching from
the "package-build" repository.

This commit does not move the library into the subdirectory.  The file
is removed without being added at the new location.  The next commit
will add the new repository as a subtree using "git subtree".

The move to a subtree is split across two commits because the second
commit as described above has to be a merge commit and one shouldn't
install any changes in a merge commit that do not come from the merged
branch or are the result of conflicts.

* Squashed 'package-build/' content from commit d70db33

git-subtree-dir: package-build
git-subtree-split: d70db33776cd52586e05e932915653ec54f41c9c