Add direnv to MELPA (#4294)
The direnv package loads environment variables from direnv into the
file-local environment variables for a buffer. This is mostly useful
when the environment variables defined in direnv control compiler or
test behaviour.

The upstream repository is at https://github.com/jml/direnv-el

I am the author and maintainer of the upstream package.