Remove character-fold+
It isn't compatible with Emacs 24.* (or older) and it is also not
compatible with Emacs 25.1 (or newer).  It is only compatible with
some development versions toward 25.1.  I have reminded Drew about
this a few months ago, but nothing has happened since.  If this
package makes a comeback, then that will likely be under the name,
char-fold+ because the library it extends has since been renamed
to char-fold.el.

From the header:
Compatibility: GNU Emacs: 25.x builds ON OR BEFORE 2015-12-10