Add el-expectations
I know we generally don't add packages from the wiki anymore, but here
I am making an exception because this package is being used by some
packages that are in Melpa and Emacsmirror and unfortunately bundle
this library.  I am asking their maintainers to remove the bundled
copy, which is harder if I cannot add "it is easily installable from
Melpa anyway".