Add recipe for pelican-mode.
pelican-mode is an Emacs minor mode for editing pages and posts in
Pelican sites. Pelican is a static site generator which can process a
variety of text file formats. For more information, see URL
https://blog.getpelican.com/.

It's intended to be used alongside a major mode for the Pelican
document. Currently supported formats are Markdown, reStructuredText,
AsciiDoc, and Org.