Update swbuff recipe
Previously the repository on the orphanage was created manually and
contained just swbuff.el.  Now it is an automated import of parts of
the upstream cvs repository.  However it now contains all files that
are located inside the same directory as that file, because cvsps
cannot process individual files (it segfaulted when I tried).