Remove bzr, cvs, darcs, fossil and svn support (#4802)
* No longer import directly from bzr, cvs, darcs, fossil or svn

The upstreams of these packages use some version control system
other than Git or Mercurial.  As you can see there are very few
such packages, which is why support was removed from Melpa (not
having to support five vcs for the benefit of a handful of
packages makes it more appealing to improve the remaining code).

See https://github.com/melpa/melpa/pull/4802.

Before this commit this was already done for these packages:

| Package (2)       | Vcs    | Last commit           | Commits last year |
|-------------------+--------+-----------------------+-------------------|
| swbuff            | cvs    | 10 years ago          |                 0 |
| vc-fossil         | fossil | 4 weeks ago           |                 1 |

Now we do it for these too:

| Package (11)      | Vcs    | Last commit           | Commits last year |
|-------------------+--------+-----------------------+-------------------|
| cg                | svn    | 7 months ago          |                 2 |
| clang-format      | svn    | 4 weeks ago           |                26 |
| color-theme       | bzr    | 8 years ago           |                 0 |
| confluence        | svn    | 6 weeks ago           |                 1 |
| darcsum           | darcs  | 3 years, 4 months ago |                 0 |
| dic-lookup-w3m    | svn    | 5 weeks ago           |                 4 |
| dsvn              | svn    | 6 weeks ago           |                 1 |
| helm-ls-svn       | svn    | 2 years, 2 months ago |                 0 |
| ruby-additional   | svn    | 3 weeks ago           |                12 |
| tex-smart-umlauts | darcs  | 1 year, 4 months ago  |                 0 |
| w3m               | cvs    | 3 weeks ago           |                74 |

... as well as `shimbun', which is part of the `w3m' repository.

A more or less up-to-date version of this table can be found at
https://emacsmirror.net/stats/melpa.html.

* Remove bzr, cvs, darcs, fossil and svn support

Not having to support five vcs for the benefit of a handful of
packages makes it more appealing to improve the remaining code).

See https://github.com/melpa/melpa/pull/4802.