Add library.el (#5228)
* Add library.el

Library.el is a PDF library manager.  Presently it only displays a
concatenated tabular list of PDF files from many locations, and allows
to open the files or view metadata from that list.  But improvements
are planned.

* Move recipes/library -> recipes/bibliothek.

Renaming the library because of the possible ambiguity the generic
name might cause.

* recipes/bibliothek: Change the names from library to bibliothek.