Various Makefile improvements (#5326)
* make: drop support for Emacsen older than 24.3

* make: suppress some irrelevant output

* make: say that we are building a package

* make: do not claim we wrote something if we didn't

If the build failed, then print nothing.

If the build succeeded either because we did a great job or because
there was nothing to be done, then print the list of relevant files
as before.  But prefix that with "Success" instead of "Wrote" because
in the later case we did not actually write those files just now.

Also put the "Success" message on a separate line so that the first
file aligns with the others.

* make: do not prefix paths with "./"

Also rearrange the variable definitions a bit.

* make: remove unnecessary let-bindings

Variable `package-build-stable' is only relevant when building a
package.  These make targets don't call any elisp that does that.
Setting that variable anyway does no harm, except that the Makefile
becomes less readable due to overlong lines.

* make: simplify setting package-build-archive-dir

* make: include config.mk if available

* make: add new LOAD_PATH variable

That makes it possible to test a development version
of `package-build' without having to first merge it
into this repository.

* make: add new LISP_CONFIG variable

* make: don't echo the commands that are being run