Remove anything and all packages that depend on it (#3004)
Anything was long ago superseded by Helm, one of the most popular
packages of all time.  Remaining users of Anything should probably
move to Helm.

Anything is one of only two packages that Melpa still fetches over
an insecure connection.  And that, combined with a maintainer who
is hard to contact, is the reason why we finally pull the plug.