Add recipe for libelcouch
The package libelcouch is an Emacs library client to communicate with
CouchDB (https://couchdb.apache.org/), a database focusing on ease of
use and having a scalable architecture.  For a user interface, please
check the elcouch project instead (which depends on this one).

https://gitlab.petton.fr/elcouch/libelcouch

I'm the package author.

No communication needed.