Upload keypress-multi-event
This package provides a method to define multiple functions to be performed for the same keypress. You can use this as a convenience feature to toggle among several functions. It was originally written as the back-end for package `home-end’, allowing those two keys to smartly cycle between moving POINT to the beginning/end of a line, the beginning/end of the window, the beginning/end of the buffer, and back to POINT.

`home-end' is being retired from debian (package `emacs-goodies') and I had been approached to maintain it independently (https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=759721#13). After four failed attempts to contact the original authors, I re-wrote the package from scratch (just try diffing!), including creating this generalized back-end that can be used for similar purposes. This very small package is a requirement of `home-end'.

My source repository: https://github.com/Boruch-Baum/MELPA_misc_contrib