Remove obsolete ./migrate script
As far as I can tell this isn't used anymore and from the age of the
script and appearance of the string "epkgs" in it I conclude that it
was once used to create Melpa recipes from a plain-text file
containing Emacsmirror package information.  That file no longer
exists (the Emacsmirror uses an SQLite database now) and as a result
this script can no longer be used to perform its intended task.