Restore quack recipe
Some user posted on an issue in the "emacsattic/quack" repository
that "quack.el" is now available over a secure connection again.
I proceeded to migrate it from the Emacsattic to the Emacsmirror,
which involves deleting the "emacsattic/quack" repository.  When
trying to say "you can submit your pr with the quack rescipe now"
I learned that the issue was actually in the "emacsattic/quack"
repository, i.e. gone.  So I am doing this myself.