Change nginx to use Debian base.
Certbot/Letsencrypt is changing up and we need a newer version of
certbot based on the emails that have been sent out. To make this happen
we are switching to a light debian Nginx image with backports enabled.
This gets us up to the correct version of certbot to avoid any
disruptions. Ran test run with the new image and everything seems fine.