Remove color-theme-solarized
This package contains an old-school theme, which requires the obsolete
"color-theme" package and a theme using the "new" theme support that
is part of Emacs itself.  The package is named after the legacy
"color-theme"-theme.

This package is not really maintained anymore.  While the last commit
is only from one and a half years ago, most issues and pull-requests
are not being responded to.

I believe users are still installing this package because it is still
being linked to from the homepage (and repository) of the original
solarized (for vim): https://ethanschoonover.com/solarized.  New users
are much better of using the `solarized-theme' package, which is well
maintained and much more popular.

The name of the library `solarized-theme.el' provided by this package
conflicts with the name of the library provided by the package by that
name.  I have asked the author to do something about it but never got
a response, see #5780.