Add json-process-client
Summary

The package json-process-client is an Emacs library to
facilitate communicating with servers that read and write JSON. The
library is responsible for starting the server and connecting with
TCP. It is also responsible for converting to and from JSON.

The library was written by Nicolas Petton in the context of Indium and
later generalized by Damien Cassou who needed something similar to
integrate basecamp within Emacs.

Repository : https://gitlab.petton.fr/nico/json-process-client

Association : With Nicolas Petton's permission, I extract part of
Indium and created json-process-client.