Update repository for helm-frame
As far as I can tell the author originally used Gitlab and then
migrated to Github without deleting the Gitlab repository or informing
us.  Then he migrated from Gitlab to self-hosting and deleted the old
Gitlab and Github repositories.

I noticed that the Gitlab repository was gone and located the web
interface of the new repository but was unable to determine an URL
than can be used to clone the it without knowing a password.

I contacted the author who informed me that he "[hasn't] fixed http
download on my new git forge yet" and then made the old Github
repository temporarily available as a temporary solution.