Get highlight from the Emacsmirror
For security reasons Melpa does not distribute any packages that
are maintained on the Emacswiki.  For this package we are forced to
make an exception because it is required by several non-emacswiki
packages, which in turn are required by Spacemacs.

The Emacsmirror imports this package like it mirrors any other
package that is maintained on the Emacswiki, by extracting the
history of the respective library and putting the result in the
`master' branch of the respective repository on the Emacsmirror;
in this case https://github.com/emacsmirror/highlight.

Melpa then mirrors the `melpa' branch of that repository.  Unlike
the `master' branch, the `melpa' branch isn't semi-automatically
updated.  Instead some user has to open a pull-request asking me to
"merge master into melpa".  I'll then review the changes and merge.
I expect this will only happen if there are breaking changes.

Closes #6436.
Re #5008.