Remove the metaweblog.el package file from the org2blog recipe.
metaweblog is a package. metaweblog.el is now stored in the org2blog repository.

metaweblog is not part of the the org2blog package though, they just live in the same place.

The section https://github.com/melpa/melpa#recipe-format explains why you need the :excludes property to handle this.