Temporarily distribute magit-section.el with magit again
The snapshot version of `magit' should not do that because it now
depends of the `magit-section' package, which consists of just this
package.

But the stable version of `magit' must contain `magit-section.el'
because at the time 2.90.1 was released the latter wasn't ready to be
distributed as a separate package.  Unfortunately Melpa decided to
recreate `magit-2.90.1' and remove `magit-section.el' retroactively
but of course without changing it to depend on `magit-section'.

The latter is a bigger issue than the former.