Remove auth-source-pass
It is part of Emacs since version 26.1 and while there still is a
separate upstream repository it is being maintained more actively
as part of Emacs than in that repository.

When it is desirable that a package that is part of Emacs is also
made available from an Elpa archive, then GNU Elpa should be used
for that purpose, not Melpa.