Add agda mode
Is part of the main agda repository: https://github.com/agda/agda
Has been used a long time but nobody bothered to add to melpa.
I wish to use nix for emacs config management and adding this
to melpa, adds it to nix.

My association is that I'm writing a blog post about agda at the moment,
but while doing that I wish to leave the state of agda
better and more accisable than I found it in.

There is an issue to do this upstream:
https://github.com/agda/agda/issues/3360
I'm doing this because it looked easy and I have reasonable
gains out of it.