pr-template: Hide instructions when no longer relevant
The pull-request instructions are placed in the text area that the
contributor uses to provide the information that those instructions
request.

Most contributors remove those instructions before submitting, as we
would like them to do, but quite a few of them do not remove this
particular instruction.

By using <!-- javascript comment syntax --> we prevent that text from
appearing in the published pull-remote description while only slightly
obfuscating it while it is being shown in the text area.