add ivy-nixos-options
This is similar to helm-nixos-options, which is in the same repo as
ivy-nixos-options and already in melpa. There are two issues here I
should point out:

- none of the files in the upstream repo have lexical binding enabled.
Since there are already recipes for all the other files in melpa
anyway, I wouldn't expect this to be a blocker. I have opened a PR
to fix this.
- This recipe doesn't build for stable melpa. The only upstream tag
is before the file was included in the repo, so make STABLE=t fails
with

Updating /home/rski/Code/melpa/working/ivy-nixos-options/ Debugger
entered--Lisp error: (error "No matching file(s) found in
/home/rski/Code/melpa...") error("No matching file(s) found in %s: %s"
"/home/rski/Code/melpa/working/ivy-nixos-options/" ("ivy-nixos-options.el"))

I can ask for a new tag upstream, but what is of most interest to me
is whether this problem is a blocker for merging the recipe or not.