org-roam: add "extensions" directory
Org-roam is undergoing a refactoring (see org-roam/org-roam#1724), which
will separate the package into 2 distinct components: the core (db,
capture etc.), and extensions which depend on the core (dailies,
graphing, protocol etc.). For clarity, these extensions will be placed
in the "extensions" directory.

This PR prepares the package for this change, including all lisp files
within the "extensions" directory.