Delete agda2-mode
`agda` and `agda-mode` are always released together and `agda-mode` needs the exactly same version of `agda` to run. For example,  a newer `agda-mode` on the `master` branch of agda/agda does not work with a stable version of `agda` from Hackage. So, `agda2-mode` simply should not be installed via `melpa`.