Add additional files to with-editor to allow building in stable.
The tagged version of with-editor and the version at HEAD have a different
directory structure. If no files are found then the build fails and currently
this is causing the stable build to fail. Temporarily adding the extra files
will solve this problem and won't hurt anything if we do end up forgetting to
clean things up.