Remove redundant :files spec from many recipes
For each of these packages its files spec resulted in the same set of
files as the default spec that is used if no explicit spec is given.

In some cases the package-specific spec was written at a time when
the default would have led to a different results.  In other cases
the additional files that were once included have since disappeared.
Yet others likely were always redundant.