Adjust recipes for new "lisp/*.el" member of default files spec
Take advantage of the addition of "lisp/*.el" to the default files
spec.  The only package for which the adjustment is mandatory is
"php-mode", as the expansion of its spec would otherwise contain
duplicates.

The "ein" repository no longer contains "lisp/zeroein.el".  The
"pdf-tools" repository no longer contains "lisp/tablist-filter.el".