Protocol buffers was moved to a new Github repo
protobuf-mode was moved to a new Github repo. You can check this by heading to https://github.com/google/protobuf and seeing that it redirects to https://github.com/protocolbuffers/protobuf .

I discovered this using straight.el and as git wasn't able to follow the redirect when attempting to install protobuf-mode.