Add recipe for new package lsp-scheme
Add support for Scheme to the lsp-mode. Currently only CHICKEN and Guile are
supported. This package relies on scheme-lsp-server available at
https://codeberg.org/rgherdt/scheme-lsp-server. I am the creator of both the
LSP server and this package.

Repository: https://codeberg.org/rgherdt/emacs-lsp-scheme

Differently than most other lsp-mode packages, lsp-scheme creates a single
instance of Scheme (Guile or CHICKEN, depending on which library is loaded) that
spawns on demand new threads providing LSP servers. This way the user can
interact with the running instance, for which a REPL is available at
*lsp-scheme*, and information is provided based on what the user defined/loaded.