No longer double down on default branch in recipes
If the branch specified by `:branch' is the same as the default branch
for the package repository, then stop specifying `:branch' because in
the absence of that setting, we default to the default branch anyway.

In some cases the specified branch wasn't just the default branch, but
the only branch.

This change comes with the risk that a package maintainer remembers
that they set `:branch' in the past and then makes another branch the
default, assuming that Melpa would continue to use the specified and
not the new default branch, which might not be appropriate for Melpa.

However, I think that risk is smaller than the risk that a maintainer
changes the name of the default branch and forgets that they have to
adjust the Melpa recipe as well because it contains a setting that is
no longer correct.

I suspect that in most cases the package maintainer wanted to use
another branch as the default when the recipe was added, but has since
changed their mind and decided to do it like almost everyone else does
it.